ehance: 将 Beat 相关的部分中间运算过程改为使用 double 以确保累计计算精度
顺便将 ReadyBeat 改为使用 uint 以简化逻辑

+++ Cyan-Stars/Assets/Scripts/Gameplay/ChartEditor/Model/ChartEditorModel.cs
@@ -76,7 +76,7 @@ namespace CyanStars.Gameplay.ChartEditor.Model
         public readonly ReactiveProperty<int> PosAccuracy = new ReactiveProperty<int>(4);
         public readonly ReactiveProperty<bool> PosMagnet = new ReactiveProperty<bool>(true);
         public readonly ReactiveProperty<int> BeatAccuracy = new ReactiveProperty<int>(2);
-        public readonly ReactiveProperty<float> BeatZoom = new ReactiveProperty<float>(1f);
+        public readonly ReactiveProperty<double> BeatZoom = new ReactiveProperty<double>(1.0d);
 
 
         /// <summary>
